CLOUDERA-BUILD. Use cdh avro version.
This updates the avro version to the internal cdh.avro.version. There
were significant issues with both avro-mapred and trevni-avro, which
have hadoop1 and hadoop2 classifiers. This project relied implicitly on
hadoop1 because Avro distributes classifer-less jars for
backward-compatibility. This commit adds avro.hadoop.classifier to the
build profiles.

Just the classifier change doesn't fix the build: the avro-mapred
dependency in trevni-avro must be excluded or else it will cause the
build to fail by pulling in the avro-mapred:hadoop1 artifact, despite
the trevni-avro dependency correctly using hadoop2. The
avro-mapred:hadoop1 artifact does not exist in internal builds.