Add asPCollection method to PTable and corresponding Conversions method (CRUNCH-483)
When using Scrunch PCollections and attempting to map to a pair of values, the
keyvalue implicit function in CanParallelDo will "upgrade" the result to a
PTable[K, V]. This is often the desired behaviour, but as Scrunch PTable is not
an extension of Scrunch PCollection, then there are cases where this is not what
is wanted. Adding asPCollection allows for an explicit "downgrade" in this cases
where it is desirable to have a PCollection[(K, V)] instead.

Signed-off-by: Josh Wills <josh.wills@gmail.com>