CRUNCH-528 Improve Pair comparison
Resolves the issue of integer overflow when comparing hash codes
of values, as well as providing stable sorting of null values
and non-equal objects with hash collisions.

Description of the underlying problem:

If the hash code values are too large, then the values will wrap
when doing subtraction. This results in a comparison function
that is not transitive.

Among other things, this makes Joins using the in-memory
pipeline not work, since the in-memory shuffler uses a TreeMap
if the key type is Comparable. Since the key in a join is a Pair
of the original key and a join tag, the key is always comparable.
With a non-transitive comparison function, it is possible for
the two join tags of the original key to sort differently,
resulting in the two join tags not being adjacent for the
original key. This results either in either the cross product
erroneously producing no values in the case of an inner join,
since the two join tags are not adjacent, or null values
appearing when they should not in the case of an outer join.